+++ src/components/sentences/VariableForm.tsx
