+++ src/components/sentences/VisualSentence.tsx
